Move pii_redactor tests to tests.rs
